Rename beta to prod, and make it happen!
This switches the URLs from beta.mybinder.org to mybinder.org!

This way, we can do a seamless switchover, start redirecting folks
from beta.mybinder.org to mybinder.org, and when the hub there is
fully drained, we can remove it too!

This also keeps the current beta.mybinder.org infrastructure alive,
in case we need to switch to it.