band-aids for false-reporting tests
Right now, we have an issue where deployments can fail (Travis is red)
misleadingly because of nginx dropping connections.

This is tricky right now, because all deployments are 'finished' when this occurs,
because nginx appears to reload configuration at runtime, not as part of any initialization.

The band-aids are twofold:

1. Wait longer after finishing deployment before running tests.
   This is because we don't seem to have a way to directly wait for nginx configuration to load.
2. Re-run tests with travis_retry if they fail.
   This accounts for potentially flaky tests, currently failing.
   With the current failure mode, trying tests again immediately after failure will always work.

These are not ideal, and should be removed if/when we can reliably wait for the nginx configuration to have reloaded.