Automatically deploy to prod if tests succeed on staging
The original reasoning for having staging and prod deployments
be both manual was we didn't have automated tests we could rely on.
So a human had to manually test staging before deploying to prod.
It was a stop-gap until we had better automation.

Now we have automated tests, and so the stop gap should go away.

This removes an entire step from our deployment workflow,
which should make it easier for a larger group of users to do.
It'll also take less time than it does now.

Someone will still need to watch for deployment failures, but since
the deployment will stop if staging deploy fails, that should be ok.

Ultimately, we need to make deployments easy enough and painless
enough that number of people doing deploys is higher and more
diverse than now, and I think this is a good step in that direction.