add service for handling user/ requests on the hub
Two containers:

- nginx handles the requests
- proxy-help monitors CHP routing table and ensures our `/users` added routes are registered

Currently, the requests simply redirect to our main binder page, which is not the best,
but is probably better than the 403 page we are serving now.
At least it alleviates the load, so it might be a good stop-gap.

In the future, this pod can serve up a more informative Binder-specific error page (or redirect to a binderhub URL for that page)