use patched grafana image
grafana's docker image is incompatible with security fixes in kubernetes 1.8.9/1.9.4

There is an open PR with a fix, but it has not been applied to the public image yet.
I applied the patch and built and pushed the image myself.
We can switch back to a public image once the patch has landed.