Switch to using a label to select node pool
This uses a label that is applied to a node pool to select where to run
the user pods. This means we can add several node pools that all have
the same label. Useful for transitioning from one pool to another or
when we want to have more than one pool for some reason.