Redirect z2jh.jupyter.org to RTD site for HTTPS
We use a cloudflare page rule to redirect z2jh.jupyter.org
to RTD. However, this doesn't work with HTTPS, since our cloudflare
HTTPS setting is set to 'Full' not 'Flexible'. 'Full' requires
the backing server (in this case mybinder.org, since that is where
the A record points to now) listen on 443 and do the redirect
appropriately. This makes it work with https too. We can remove
the cloudflare page rule in the future if we want to.