Update secrets/ban.py to match master branch
I had a merge conflict on my federation pull request that was
difficult to resolve because VSCode couldn't render the diff,
I assume because of the encryption. So what I'm doing in this commit
is copying the file from git pull master and pasting it into the
file on my branch and hoping that resolves the conflict.