Switch to using absolute disk usage for DIND on OVH
On the OVH cluster the DIND and K8S dockerds both use the same partition
to store images. As both check the relative usage of the partition to
decide when to act the DIND GC job ends up deleting all images everytime
it runs because the disk is too full (and just removing the DIND images
doesn't drop it far enough). This switches it to use a size in GB as
limit. This should balance the disk usage between the two dockerds.