Tell helm to clean up on failed deploys
This fixes a large class of problems where deploying (again) fails. The
problem is that a previous deploy failed and left tiller thinking that
the cluster is in a different state than it really is. Adding the
cleanup-on-fail flag tells helm to attempt to clean up. It is not
perfect but better than nothing.