Stop deploying to OVH
The OVH cluster doesn't work right now (certs and etcd storage). The failing
deploy to it means we never run the tests on GKE or Turing. Instead we are
getting used to the deploy being red. As a result this removes the deploy and test
of OVH until we can get it fixed up.