k8s 1.16 compatibility
Various apiVersion for Deployment resources became unsupported in k8s
1.16, so this commit makes us use apps/v1 instead. While doing so we
also need to add the `selector` field of the Deployment specification.

Based on Helm's recommended practices, I exclude the heritage label from
the selector, but I don't update other deployments with this practice as
it is not something I think will matter.