Configure Docker Hub credentials for OVH
Use credentials when pulling from docker hub in order to get a higher
rate limit. So far the OVH deployment still had credentials configured
for the internal docker registry whcih we stopped using a while ago.
This changes the credentials to the docker hub ones which we have been
using for pushing images already.