Test build with cached-minimal-dockerfile repo instead of requirements repo
The cached-minimal-dockerfile is a repo that is optimized for build
speed and can reduce time to the absolute minimum by making repo2docker
build a Dockerfile with a single FROM statement downloading a pre-build
image of ~50 MB or so.

The benefit of this change will only be seen whenever the test_build job
is triggering a build to run on a node without cached results, but at
that point, it will make a very significant difference.

For more information about this ci optimized repo, see its
[README.md](https://github.com/binderhub-ci-repos/cached-minimal-dockerfile).
The repo is currently in use also by jupyterhub/binderhub's CI system.

Note that as part of making this change, I created a new deployment key
as github doesn't allow for a deployment key to be reused for two repo's,
and the old key was used for binderhub-ci-repos/requirements already.