federation-redirect: check versions before health
checking versions on the prime host updates the version to be checked against

if the prime host is unhealthy (e.g. full), the version will not be checked
and therefore the prime version will not be updated.

This can result in starving fully up-to-date federation members as long as the prime version is full
(which will be more likely to stay that way, as we don't stop routing requests to prime when it is full)