Add a 2i2c federation member on Hetzner
https://2i2c.mybinder.org/ is up now!

https://github.com/2i2c-org/2i2c-org.github.io/pull/356/files#diff-7244b57e647732dd6a8f006bdf63943e1dcb813fa1a085073522ccf40e2cdfc6
has more context - that's also an announcement blog post. It came
together quickly.

This is a single node k3s cluster running on Hetzner. It's not as
large as we'd like it to be - which is CCX63 on https://www.hetzner.com/cloud.
That's 48 vCPUs and 192GB of RAM. And with k3s, we can override the
number of pods on a node. Given the current guarantee of 450M, we can
put approximately 400 pods on this one node! That runs out to less
than $1 / month per user capacity which is pretty good.

Still need to figure out:

1. Access for everyone else on the team
2. Resize the server to be big, and set up k3s again there from
   scratch + document (I simply followed the quickstart with traefik
   disabled)
3. Test prometheus and grafana
4. Add 2i2c to list of supporters

Am excited to try this out and see how it goes.

Thanks to @choldgraf, @colliand, @jmunroe and others at 2i2c
for supporting me through this.