Add another smaller 2i2c federation member + k3s docs
I wanted to understand what it would take to run a federation
member on a small machine that does *not* have object storage
nearby (I'm trying to get some new members where this is true).
I also wanted to run through adding a new k3s member one more time
so I can document it appropriately

Sooooooo I bought a small server (16 cores AMD Ryzen with 64G of RAM
and 1TB RAID1 SSD) via the wonderful [hetzner auction](https://www.hetzner.com/sb/)
system. They sell older systems that are decomissioned from 'production'
users but those are perfect for us. This costs 2i2c approximately
60$ a month in costs. If the server dies (because this is a physical
object that has physical faults), we can just stop paying for it
or try to ressurect it as we wish.

I installed Ubuntu 24.04 and ran through k3s setup again, documenting
every step I did! This helps future people joining the federation this
way.

Since this is in the same datacenter as the regular bigger 2i2c
federation member, we can reuse the same object store backend for
the registry! I can experiment with using filesystem for the registry
in a future separate commit.

I've added this to the federation with a 5% weight so it gets a little
traffic but not much.

Thanks to 2i2c for sponsoring this!