Use openebs localpath provisioner for enforcing filesystem size limits
If we want to run a federation member without access to object storage,
we need:

1. To enforce a filesystem limit for how much the registry can use,
   as otherwise it will simply fill up the whole system
2. Enforce a filesystem limit for how much dind can use for same reason

The default k3s localpath provisioner doesn't support quotas or
size limits. However, https://github.com/openebs/dynamic-localpv-provisioner/
does! This deploys it, and uses it to run the registry off local storage,
rather than object storage!

We still need a generic registry cleaner and to enforce limits on the dind
pod size, but this is a step in the right direction.