Increase and centralize build memory limits
Follow-up to https://github.com/jupyterhub/mybinder.org-deploy/pull/3204,
which failed because it removed the limit, which caused it to be
lower than the request. All the reasons given in that PR still apply.

Here I bump up the limit enough until the super popular
https://github.com/jupyterlab/jupyterlab-demo is able to build.
It's not a functionally super complex environment.yml, and even with
simplifications in https://github.com/jupyterlab/jupyterlab-demo/pull/147,
it fails to build without the memory increase limit. As mentioned in
the other PR, https://github.com/jupyterhub/repo2docker/pull/1402 will
get rid of this kind of limit anyway, so it's ok to raise this.