hetzner: Run imagecleaner more aggressively
The reason we had to revert https://github.com/jupyterhub/mybinder.org-deploy/pull/3228
was that the image cache was growing big, and so was the buildkit
cache. So this meant images were being stored in *three* places:

1. the buildkit cache (dind)
2. the image cache (dind)
3. the containerd store (by kubernetes, when it pulls)

1 and 2 were both set to about 300, so this caused the disk to
fill up.

Here, I more aggressively prune the image cache, as it is simply
not used when buildkit is used. Once I verify this works ok, I'll
turn on buildkit building again and we see how it goes.