Make GESIS node use its own registry
There was a fully configured image registry on the GESIS node
but it was not being used. The object storage backend is shared,
but the registry itself is not shared - it's stateless, and needs
its own server to run. This is primarily to prevent the situation
where the 2i2c server being down causes issues elsewhere - this way
the only dependency is on the object storage itself.