Use different http.secret for each registry
Potentially, the issue being reported in
https://jupyter.zulipchat.com/#narrow/channel/469744-jupyterhub/topic/Building.20on.20MyBinder.20okay.20for.20others.3F/with/535003764
is due to the two registries pushing to the same object storage at
the same time, with the same hmac key, and overwriting each other. I
don't have a clear way to verify this, but let's try this out. The
two registries are *not* behind a load balancer, so perhaps they
need to be in different object stores.